Acertando o tipo de retorno para uint8_t que é o tipo do ID de tabela.

+++ scratch/svelte/infrastructure/backhaul-controller.h
@@ -141,7 +141,7 @@ protected:
    * \param slice The slice ID.
    * \return The pipeline table for this slice.
    */
-  uint16_t GetSliceTable (SliceId slice) const;
+  uint8_t GetSliceTable (SliceId slice) const;
 
   /**
    * Notify this controller of a new bearer context created.
